backend/rag-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-storage==2.14.0
langchain==0.0.350
langchain-community==0.0.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0

# Required for main.py
langchain-pinecone==0.2.13
langchain-openai==0.2.8
pinecone==5.4.0
pydantic==2.9.2
